+++ lib/joystick/joystick_move_to_position.dart
@@ -28,7 +28,7 @@ class JoystickMoveToPosition extends JoystickController {
     if (_pointer == event.pointer) {
       if (!_interfaceReceiveInteraction) {
         final absolutePosition =
-            this.gameRef.camera.screenPositionToWorld(event.position);
+            this.gameRef.screenPositionToWorld(event.position);
         moveTo(absolutePosition.toVector2());
       }
     }