+++ lib/util/gestures/gestures.dart
@@ -1,2 +1,3 @@
 export 'package:bonfire/util/gestures/drag_gesture.dart';
+export 'package:bonfire/util/gestures/mouse_hover_gesture.dart';
 export 'package:bonfire/util/gestures/tap_gesture.dart';