+++ lib/util/gestures/mouse_hover_gesture.dart
@@ -1,0 +1,30 @@
+import 'package:bonfire/base/game_component.dart';
+import 'package:flutter/gestures.dart';
+
+mixin MouseHoverGesture on GameComponent {
+  bool enableHover = true;
+  @override
+  void handlerPointerHover(PointerHoverEvent event) {
+    int pointer = event.pointer;
+    Offset position = event.localPosition;
+    onHoverScreen(pointer, position);
+    if (this.isHud) {
+      if (this.position.contains(position)) {
+        onHoverIn(pointer, position);
+      } else {
+        onHoverOut(pointer, position);
+      }
+    } else {
+      final absolutePosition = this.gameRef.screenPositionToWorld(position);
+      if (this.position.contains(absolutePosition)) {
+        onHoverIn(pointer, position);
+      } else {
+        onHoverOut(pointer, position);
+      }
+    }
+  }
+
+  void onHoverScreen(int pointer, Offset position);
+  void onHoverIn(int pointer, Offset position);
+  void onHoverOut(int pointer, Offset position);
+}