+++ lib/util/mixins/pointer_detector.dart
@@ -1,4 +1,5 @@
 import 'package:bonfire/util/gestures/drag_gesture.dart';
+import 'package:bonfire/util/gestures/mouse_hover_gesture.dart';
 import 'package:bonfire/util/gestures/tap_gesture.dart';
 import 'package:flutter/gestures.dart';
 
@@ -7,6 +8,7 @@ mixin PointerDetector {
   void onPointerMove(PointerMoveEvent event) {}
   void onPointerUp(PointerUpEvent event) {}
   void onPointerCancel(PointerCancelEvent event) {}
+  void onPointerHover(PointerHoverEvent event) {}
 }
 
 abstract class PointerDetectorHandler {
@@ -14,10 +16,13 @@ abstract class PointerDetectorHandler {
   void handlerPointerMove(PointerMoveEvent event) {}
   void handlerPointerUp(PointerUpEvent event) {}
   void handlerPointerCancel(PointerCancelEvent event) {}
+  void handlerPointerHover(PointerHoverEvent event) {}
 
   bool hasGesture() {
     if (this is DragGesture && (this as DragGesture).enableDrag) return true;
     if (this is TapGesture && (this as TapGesture).enableTab) return true;
+    if (this is MouseHoverGesture && (this as MouseHoverGesture).enableHover)
+      return true;
     return false;
   }
 }